requirements.txt
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.9.0
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
Faker==37.1.0
Flask==3.1.1
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.1.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
praw==7.7.1
prawcore==2.4.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyzmq==26.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
